+++ src/main/org/firebirdsql/gds/ng/AbstractFbStatement.java
@@ -655,4 +655,22 @@ public abstract class AbstractFbStatement implements FbStatement {
         RowDescriptor fieldDescriptor = getRowDescriptor();
         return fieldDescriptor != null && fieldDescriptor.getCount() > 0;
     }
+
+    /**
+     * Signals the start of an execute for this statement.
+     *
+     * @return {@code OperationCloseHandle} handle for the operation
+     */
+    protected final OperationCloseHandle signalExecute() {
+        return FbDatabaseOperation.signalExecute(getDatabase());
+    }
+
+    /**
+     * Signals the start of a fetch for this statement.
+     *
+     * @return {@code OperationCloseHandle} handle for the operation
+     */
+    protected final OperationCloseHandle signalFetch() {
+        return FbDatabaseOperation.signalFetch(getDatabase());
+    }
 }