+++ src/main/org/firebirdsql/jdbc/AbstractFieldMetaData.java
@@ -114,6 +114,7 @@ public abstract class AbstractFieldMetaData implements Wrapper {
         case ISCConstants.SQL_INT64:
         case ISCConstants.SQL_DEC16:
         case ISCConstants.SQL_DEC34:
+        case ISCConstants.SQL_INT128:
         case ISCConstants.SQL_DEC_FIXED:
             return true;
         default:
@@ -244,6 +245,7 @@ public abstract class AbstractFieldMetaData implements Wrapper {
         case ISCConstants.SQL_DEC16:
         case ISCConstants.SQL_DEC34:
             return "DECFLOAT";
+        case ISCConstants.SQL_INT128:
         case ISCConstants.SQL_DEC_FIXED:
             if (sqlSubtype == SUBTYPE_NUMERIC) {
                 return "NUMERIC";
@@ -397,6 +399,8 @@ public abstract class AbstractFieldMetaData implements Wrapper {
         case ISCConstants.SQL_DEC34:
         case ISCConstants.SQL_DEC_FIXED:
             return 34;
+        case ISCConstants.SQL_INT128:
+            return 38;
         default:
             return 0;
         }